Fixed param size in stack frame. Removed some leftovers (old TryFrame)
git-svn-id: svn://falconpl.org/falcon/core/trunk@270 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ include/falcon/stackframe.h
@@ -49,7 +49,7 @@ typedef struct tag_StackFrame
    bool m_initFrame;
    bool m_break;
    bool m_suspend;
-   byte m_param_count;
+   uint16 m_param_count;
    uint32 m_stack_base;
    uint32 m_try_base;
    Item m_sender;
